lipcoding-competition/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
sqlalchemy
